Fix: Decimal타입을 모두 Long으로 변경

+++ backend/src/main/java/com/ssafy/keeping/domain/charge/dto/response/CancelListResponseDto.java
@@ -13,7 +13,7 @@ public class CancelListResponseDto {
     
     private String transactionUniqueNo;
     private String storeName;
-    private BigDecimal paymentAmount;
+    private Long paymentAmount;
     private LocalDateTime transactionTime;
-    private BigDecimal remainingBalance;
+    private Long remainingBalance;
 }
